requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.6.0
black==22.6.0
click==8.1.3
coverage==6.4.2
Django==4.0.6
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
flake8-isort==4.2.0
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
whitenoise==6.2.0
